sdk/js: bump injective sdk
