+++ sdk/js/src/token_bridge/createWrapped.ts
@@ -6,27 +6,22 @@ import {
   Transaction,
 } from "@solana/web3.js";
 import { MsgExecuteContract } from "@terra-money/terra.js";
+import { MsgExecuteContract as XplaMsgExecuteContract } from "@xpla/xpla.js";
 import { Algodv2 } from "algosdk";
 import { Types } from "aptos";
 import BN from "bn.js";
-import { ethers, Overrides } from "ethers";
+import { Overrides, ethers } from "ethers";
 import { fromUint8Array } from "js-base64";
-import {
-  TransactionSignerPair,
-  _parseVAAAlgorand,
-  _submitVAAAlgorand,
-} from "../algorand";
-import { Bridge__factory } from "../ethers-contracts";
-import { submitVAAOnInjective } from "./redeem";
 import { FunctionCallOptions } from "near-api-js/lib/account";
 import { Provider } from "near-api-js/lib/providers";
-import { callFunctionNear } from "../utils";
-import { MsgExecuteContract as XplaMsgExecuteContract } from "@xpla/xpla.js";
+import { TransactionSignerPair, _submitVAAAlgorand } from "../algorand";
 import {
   createWrappedCoin as createWrappedCoinAptos,
   createWrappedCoinType as createWrappedCoinTypeAptos,
 } from "../aptos";
+import { Bridge__factory } from "../ethers-contracts";
 import { createCreateWrappedInstruction } from "../solana/tokenBridge";
+import { callFunctionNear } from "../utils";
 import { SignedVaa } from "../vaa";
 
 export async function createWrappedOnEth(
@@ -52,8 +47,6 @@ export async function createWrappedOnTerra(
     },
   });
 }
-
-export const createWrappedOnInjective = submitVAAOnInjective;
 
 export function createWrappedOnXpla(
   tokenBridgeAddress: string,
@@ -132,9 +125,9 @@ export async function createWrappedOnNear(
 /**
  * Constructs payload to create wrapped asset type. The type is of form `{{address}}::coin::T`,
  * where address is `sha256_hash(tokenBridgeAddress | chainID | "::" | originAddress | 0xFF)`.
- * 
- * Note that the typical createWrapped call is broken into two parts on Aptos because we must first 
- * create the CoinType that is used by `create_wrapped_coin<CoinType>`. Since it's not possible to 
+ *
+ * Note that the typical createWrapped call is broken into two parts on Aptos because we must first
+ * create the CoinType that is used by `create_wrapped_coin<CoinType>`. Since it's not possible to
  * create a resource and use it in the same transaction, this is broken into separate transactions.
  * @param tokenBridgeAddress Address of token bridge
  * @param attestVAA Bytes of attest VAA
@@ -148,11 +141,11 @@ export function createWrappedTypeOnAptos(
 }
 
 /**
- * Constructs payload to create wrapped asset. 
- * 
- * Note that this function is typically called in tandem with `createWrappedTypeOnAptos` because 
- * we must first create the CoinType that is used by `create_wrapped_coin<CoinType>`. Since it's 
- * not possible to create a resource and use it in the same transaction, this is broken into 
+ * Constructs payload to create wrapped asset.
+ *
+ * Note that this function is typically called in tandem with `createWrappedTypeOnAptos` because
+ * we must first create the CoinType that is used by `create_wrapped_coin<CoinType>`. Since it's
+ * not possible to create a resource and use it in the same transaction, this is broken into
  * separate transactions.
  * @param tokenBridgeAddress Address of token bridge
  * @param attestVAA Bytes of attest VAA
