+++ sdk/js/src/token_bridge/getIsTransferCompleted.ts
@@ -1,4 +1,3 @@
-import { ChainGrpcWasmApi } from "@injectivelabs/sdk-ts";
 import { Commitment, Connection, PublicKeyInitData } from "@solana/web3.js";
 import { LCDClient } from "@terra-money/terra.js";
 import { LCDClient as XplaLCDClient } from "@xpla/xpla.js";
@@ -9,16 +8,12 @@ import { ethers } from "ethers";
 import { fromUint8Array } from "js-base64";
 import { Provider } from "near-api-js/lib/providers";
 import { redeemOnTerra } from ".";
-import {
-  ensureHexPrefix,
-  parseSmartContractStateResponse,
-  TERRA_REDEEMED_CHECK_WALLET_ADDRESS,
-} from "..";
+import { TERRA_REDEEMED_CHECK_WALLET_ADDRESS, ensureHexPrefix } from "..";
 import {
   BITS_PER_KEY,
-  calcLogicSigAccount,
   MAX_BITS,
   _parseVAAAlgorand,
+  calcLogicSigAccount,
 } from "../algorand";
 import { TokenBridgeState } from "../aptos/types";
 import { getSignedVAAHash } from "../bridge";
@@ -26,7 +21,7 @@ import { Bridge__factory } from "../ethers-contracts";
 import { getClaim } from "../solana/wormhole";
 import { safeBigIntToNumber } from "../utils/bigint";
 import { callFunctionNear } from "../utils/near";
-import { parseVaa, SignedVaa } from "../vaa/wormhole";
+import { SignedVaa, parseVaa } from "../vaa/wormhole";
 
 export async function getIsTransferCompletedEth(
   tokenBridgeAddress: string,
@@ -105,32 +100,6 @@ export async function getIsTransferCompletedTerra2(
     }
   );
   return result.is_redeemed;
-}
-
-/**
- * Return if the VAA has been redeemed or not
- * @param tokenBridgeAddress The Injective token bridge contract address
- * @param signedVAA The signed VAA byte array
- * @param client Holds the wallet and signing information
- * @returns true if the VAA has been redeemed.
- */
-export async function getIsTransferCompletedInjective(
-  tokenBridgeAddress: string,
-  signedVAA: Uint8Array,
-  client: ChainGrpcWasmApi
-): Promise<boolean> {
-  const queryResult = await client.fetchSmartContractState(
-    tokenBridgeAddress,
-    Buffer.from(
-      JSON.stringify({
-        is_vaa_redeemed: {
-          vaa: fromUint8Array(signedVAA),
-        },
-      })
-    ).toString("base64")
-  );
-  const parsed = parseSmartContractStateResponse(queryResult);
-  return parsed.is_redeemed;
 }
 
 export async function getIsTransferCompletedXpla(
