+++ sdk/js/src/token_bridge/getIsWrappedAsset.ts
@@ -1,27 +1,12 @@
-import { ChainGrpcWasmApi } from "@injectivelabs/sdk-ts";
-import {
-  Commitment,
-  Connection,
-  PublicKey,
-  PublicKeyInitData,
-} from "@solana/web3.js";
+import { Commitment, Connection, PublicKeyInitData } from "@solana/web3.js";
 import { LCDClient } from "@terra-money/terra.js";
 import { Algodv2, getApplicationAddress } from "algosdk";
 import { AptosClient } from "aptos";
 import { ethers } from "ethers";
 import { Bridge__factory } from "../ethers-contracts";
-import {
-  CHAIN_ID_INJECTIVE,
-  ensureHexPrefix,
-  coalesceModuleAddress,
-  tryNativeToHexString,
-} from "../utils";
-import { deriveWrappedMetaKey, getWrappedMeta } from "../solana/tokenBridge";
+import { getWrappedMeta } from "../solana/tokenBridge";
+import { coalesceModuleAddress, ensureHexPrefix } from "../utils";
 import { safeBigIntToNumber } from "../utils/bigint";
-import {
-  getForeignAssetSolana,
-  getForeignAssetInjective,
-} from "./getForeignAsset";
 
 /**
  * Returns whether or not an asset address on Ethereum is a wormhole wrapped asset
@@ -47,31 +32,6 @@ export async function getIsWrappedAssetTerra(
   assetAddress: string
 ): Promise<boolean> {
   return false;
-}
-
-/**
- * Checks if the asset is a wrapped asset
- * @param tokenBridgeAddress The address of the Injective token bridge contract
- * @param client Connection/wallet information
- * @param assetAddress Address of the asset in Injective format
- * @returns true if asset is a wormhole wrapped asset
- */
-export async function getIsWrappedAssetInjective(
-  tokenBridgeAddress: string,
-  client: ChainGrpcWasmApi,
-  assetAddress: string
-): Promise<boolean> {
-  const hexified = tryNativeToHexString(assetAddress, "injective");
-  const result = await getForeignAssetInjective(
-    tokenBridgeAddress,
-    client,
-    CHAIN_ID_INJECTIVE,
-    new Uint8Array(Buffer.from(hexified))
-  );
-  if (result === null) {
-    return false;
-  }
-  return true;
 }
 
 /**
